fix CONTEXT64 EFlags size and child process PEB in thread context
CONTEXT64.EFlags was c_uint64 (8 bytes) but the real Windows CONTEXT64
has DWORD EFlags (4 bytes). This shifted all subsequent fields (Rax,
Rcx, Rdx, etc.) by 8 bytes, so malware reading Rdx at the correct
offset 0x88 would get zero instead of the PEB pointer.

Also initialize the child process thread context in create_process():
set Rdx (AMD64) or Ebx (x86) to the child PEB address and populate
PEB.ImageBaseAddress. This allows process hollowing malware to find
and update the remote PEB via GetThreadContext/WriteProcessMemory.

+++ speakeasy/winenv/defs/windows/windows.py
@@ -227,7 +227,7 @@ class CONTEXT64(EmuStruct):
         self.SegFs = ct.c_uint16
         self.SegGs = ct.c_uint16
         self.SegSs = ct.c_uint16
-        self.EFlags = ct.c_uint64
+        self.EFlags = ct.c_uint32
 
         self.Dr0 = ct.c_uint64
         self.Dr1 = ct.c_uint64
